+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Keep an eye on this section for release updates and changelogs.
 - Displaying finished and errornous tasks count as a badge number
 - Minor UI polish
 - gpt-3.5-turbo-16k support now is internalised and minions choose when to use it.
+- Improved and simplified setup
+- Monitoring and automatically running tasks for comments tagged with "TODO:" (Tag is configurable in the settings)
 
 ## [0.9.8]
 - Copy adjustments